Moving tag static component into specific class component, generic components file no longer needed
